pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "healthcare-sim-sdk"
version = "0.1.0"
description = "Healthcare Intervention Simulation SDK for AI deployment evaluation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Michael Draugelis"},
]
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
    "jupyter>=1.0",
    "matplotlib>=3.5",
    "nbconvert>=7.0",
    "pyyaml>=6.0",
    # Statistical dependencies for monitoring validation (tiered dashboard)
    "statsmodels>=0.14.0",
    "scikit-learn>=1.3.0",
    "hydra-joblib-launcher>=1.2.0",
]

[tool.setuptools.packages.find]
include = ["healthcare_sim_sdk*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.flake8]
max-line-length = 88
